+++ 19_CDIO_03/src/boundaries/GUI.java
@@ -1,0 +1,5 @@
+package boundaries;
+
+public class GUI {
+
+}
