+++ 19_CDIO_03/src/controller/GameController.java
@@ -1,4 +1,4 @@
-package spil;
+package controller;
 
 /**
  *
