+++ 19_CDIO_03/src/entity/Cup.java
@@ -1,4 +1,4 @@
-package spil;
+package entity;
 
 public class Cup {
