+++ 19_CDIO_03/src/entity/Dice.java
@@ -1,4 +1,4 @@
-package spil;
+package entity;
 
 public class Dice {
 	private int facevalue;
