+++ 19_CDIO_03/src/entity/Fleet.java
@@ -1,4 +1,4 @@
-package spil;
+package entity;
 
 public class Fleet extends Ownable {
