+++ 19_CDIO_03/src/entity/LaborCamp.java
@@ -1,4 +1,4 @@
-package spil;
+package entity;
 
 public class LaborCamp extends Ownable {
