+++ 19_CDIO_03/src/entity/Ownable.java
@@ -1,4 +1,4 @@
-package spil;
+package entity;
 
 public class Ownable extends Field {
