+++ 19_CDIO_03/src/entity/Refuge.java
@@ -1,4 +1,4 @@
-package spil;
+package entity;
 
 public class Refuge extends Field {
