+++ 19_CDIO_03/src/entity/Tax.java
@@ -1,4 +1,4 @@
-package spil;
+package entity;
 
 public class Tax extends Field {
